Add client-python to supported apps
Client is in development and soon to be released out of alpha, but it is supported. I would like to add it to this list so it gets more attraction.
I also want this to be updated at main kubernetes repo, how does that work from this repo to main repo? ref: kubernetes/kubernetes#37719